requirements.txt
discord.py==2.0.1
jishaku==2.5.1
aiosqlite==0.17.0
orjson==3.8.0